+++ src/Public/ConvertTo-String.ps1
@@ -46,6 +46,11 @@ function ConvertTo-String {
         [Parameter(ValueFromPipelineByPropertyName, ParameterSetName = 'flowchartSubgraph')]
         [string] $Title,
 
+        # Configuration of the diagram.
+        [Parameter(ValueFromPipelineByPropertyName, ParameterSetName = 'erDiagram')]
+        [Parameter(ValueFromPipelineByPropertyName, ParameterSetName = 'flowchart')]
+        [hashtable] $Config,
+
         #endregion
 
         #region erDiagram
@@ -286,9 +291,16 @@ function ConvertTo-String {
                     $Relations | ConvertTo-String | Write-Output
                 }
                 flowchart {
-                    if ( $Title ) {
+                    if ( $Title -or $Config ) {
                         '---' | Write-Output
-                        "title: $Title" | Write-Output
+                        $frontmatter = [ordered]@{}
+                        if ( $Title ) {
+                            $frontmatter.title = $Title
+                        }
+                        if ( $Config ) {
+                            $frontmatter.config = $Config
+                        }
+                        $frontmatter | ConvertTo-Yaml | Write-Output
                         '---' | Write-Output
                     }
                     switch ( $Orientation ) {
@@ -303,8 +315,8 @@ function ConvertTo-String {
                         $Classes | ConvertTo-String -FromFlowchartClass | Write-Output
                         $Nodes | ConvertTo-String -FromFlowchartNode | Write-Output
                         $Clicks | ConvertTo-String -FromFlowchartClick | Write-Output
-                        $Links | ConvertTo-String -FromFlowchartLink | Write-Output
                         ( $Subgraphs | ConvertTo-String -FromFlowchartSubgraph ) -split [Environment]::NewLine | Write-Output
+                        $Links | ConvertTo-String -FromFlowchartLink | Write-Output
                     ) | ForEach-Object { "    $_" | Write-Output }
                 }
                 C4ComponentDiagram {
@@ -323,6 +335,7 @@ function ConvertTo-String {
                     ) | ForEach-Object { "    $_" | Write-Output }
                 }
                 flowchartLink {
+                    $escapedText = [System.Web.HttpUtility]::HtmlEncode($Text)
                     Write-Output "$SourceNode $(
                         switch ( $Line ) {
                             solid { "$(
@@ -396,26 +409,27 @@ function ConvertTo-String {
                                 Write-Error "convert $_ is not supported."
                             }
                         }
-                    )$( if ( $Text ) { "|$Text|" } ) $DestinationNode"
+                    )$( if ( $Text ) { "|$escapedText|" } ) $DestinationNode"
                 }
                 flowchartNode {
+                    $escapedText = [System.Web.HttpUtility]::HtmlEncode($Text ? $Text : $Key)
                     Write-Output "$Key$(
                         switch ( $Shape ) {
-                            '' { "$( $Text ? "[$Text]" : '' )" }
-                            rectangle { "[$( $Text ? $Text : $Key )]" }
-                            round-edges { "($( $Text ? $Text : $Key ))" }
-                            stadium { "([$( $Text ? $Text : $Key )])" }
-                            subroutine { "[[$( $Text ? $Text : $Key )]]" }
-                            cylindrical { "[($( $Text ? $Text : $Key ))]" }
-                            circle { "(($( $Text ? $Text : $Key )))" }
-                            asymmetric { ">$( $Text ? $Text : $Key )]" }
-                            rhombus { "{$( $Text ? $Text : $Key )}" }
-                            hexagon { "{{$( $Text ? $Text : $Key )}}" }
-                            parallelogram { "[/$( $Text ? $Text : $Key )/]" }
-                            parallelogram-alt { "[\$( $Text ? $Text : $Key )\]" }
-                            trapezoid { "[/$( $Text ? $Text : $Key )\]" }
-                            trapezoid-alt { "[\$( $Text ? $Text : $Key )/]" }
-                            double-circle { "((($( $Text ? $Text : $Key ))))" }
+                            '' { "$( $Text ? "[$escapedText]" : '' )" }
+                            rectangle { "[$escapedText]" }
+                            round-edges { "($escapedText)" }
+                            stadium { "([$escapedText])" }
+                            subroutine { "[[$escapedText]]" }
+                            cylindrical { "[($escapedText)]" }
+                            circle { "(($escapedText))" }
+                            asymmetric { ">$escapedText]" }
+                            rhombus { "{$escapedText}" }
+                            hexagon { "{{$escapedText}}" }
+                            parallelogram { "[/$escapedText/]" }
+                            parallelogram-alt { "[\$escapedText\]" }
+                            trapezoid { "[/$escapedText\]" }
+                            trapezoid-alt { "[\$escapedText/]" }
+                            double-circle { "((($escapedText)))" }
                             Default {
                                 Write-Error "'$_' is not supported for Node Shape."
                             }
@@ -425,15 +439,15 @@ function ConvertTo-String {
                     Write-Output "classDef $Name $Style"
                 }
                 flowchartClick {
-                    Write-Output "click $Node ""$Url""$( if ( $Tooltip ) {  ' "' + $Tooltip + '"' } )$( if ( $Target ) { " _$Target" } )"
+                    Write-Output "click $Node ""$Url""$( if ( $Tooltip ) {  ' "' + [System.Web.HttpUtility]::HtmlEncode($Tooltip) + '"' } )$( if ( $Target ) { " _$Target" } )"
                 }
                 flowchartSubgraph {
-                    Write-Output "subgraph $Key$( $Title ? " [$Title]" : '' )"
+                    Write-Output "subgraph $Key$( $Title ? ' [' + [System.Web.HttpUtility]::HtmlEncode($Title) + ']' : '' )"
                     $(
                         $Nodes | ConvertTo-String -FromFlowchartNode | Write-Output
                         $Clicks | ConvertTo-String -FromFlowchartClick | Write-Output
-                        $Links | ConvertTo-String -FromFlowchartLink | Write-Output
                         ( $Subgraphs | ConvertTo-String -FromFlowchartSubgraph ) -split [Environment]::NewLine | Write-Output
+                        $Links | ConvertTo-String -FromFlowchartLink | Write-Output
                     ) | ForEach-Object { "    $_" | Write-Output }
                     Write-Output "end"
                 }