+++ test/ConvertTo-String.Tests.ps1
@@ -108,6 +108,44 @@ flowchart
             }
         }
 
+        Context with-title {
+
+            BeforeEach {
+                $diagram = New-MermaidDiagram -Flowchart -Orientation left-to-right -Title 'Test Diagram'
+            }
+
+            It works-empty {
+                $output = $diagram | ConvertTo-MermaidString -ErrorAction Stop
+                $output | Should -Not -BeNullOrEmpty
+                $output | Should -Be (@"
+---
+title: Test Diagram
+---
+flowchart LR
+"@.Replace("`r`n", [Environment]::NewLine))
+            }
+        }
+
+        Context with-config {
+
+            BeforeEach {
+                $diagram = New-MermaidDiagram -Flowchart -Config @{ flowchart = @{ defaultRenderer = 'elk' } }
+            }
+
+            It works-empty {
+                $output = $diagram | ConvertTo-MermaidString -ErrorAction Stop
+                $output | Should -Not -BeNullOrEmpty
+                $output | Should -Be (@"
+---
+config:
+  flowchart:
+    defaultRenderer: elk
+---
+flowchart
+"@.Replace("`r`n", [Environment]::NewLine))
+            }
+        }
+
         Context with-orientation {
 
             BeforeEach {
@@ -120,24 +158,6 @@ flowchart
                 $output | Should -Be (@"
 flowchart LR
 "@.Replace("`r`n", [Environment]::NewLine))
-            }
-
-            Context with-title {
-
-                BeforeEach {
-                    $diagram = New-MermaidDiagram -Flowchart -Orientation left-to-right -Title 'Test Diagram'
-                }
-
-                It works-empty {
-                    $output = $diagram | ConvertTo-MermaidString -ErrorAction Stop
-                    $output | Should -Not -BeNullOrEmpty
-                    $output | Should -Be (@"
----
-title: Test Diagram
----
-flowchart LR
-"@.Replace("`r`n", [Environment]::NewLine))
-                }
             }
 
             Context nodes {
@@ -160,7 +180,7 @@ flowchart LR
 
                 Context no-shape {
                     BeforeEach {
-                        $diagram | Add-MermaidFlowchartNode A node
+                        $diagram | Add-MermaidFlowchartNode A 'node "'
                     }
 
                     It works {
@@ -168,7 +188,7 @@ flowchart LR
                         $output | Should -Not -BeNullOrEmpty
                         $output | Should -Be (@"
 flowchart LR
-    A[node]
+    A[node &quot;]
 "@.Replace("`r`n", [Environment]::NewLine))
                     }
 
@@ -440,7 +460,7 @@ flowchart LR
                 Context Text-on-links {
 
                     BeforeEach {
-                        $diagram | Add-MermaidFlowchartLink A B -Text 'text' -DestinationHead open
+                        $diagram | Add-MermaidFlowchartLink A B -Text 'text "' -DestinationHead open
                     }
 
                     It works {
@@ -448,7 +468,7 @@ flowchart LR
                         $output | Should -Not -BeNullOrEmpty
                         $output | Should -Be (@"
 flowchart LR
-    A ---|text| B
+    A ---|text &quot;| B
 "@.Replace("`r`n", [Environment]::NewLine))
                     }
                 }
@@ -456,7 +476,7 @@ flowchart LR
                 Context A-link-with-arrow-head-and-text {
 
                     BeforeEach {
-                        $diagram | Add-MermaidFlowchartLink A B -Text 'text'
+                        $diagram | Add-MermaidFlowchartLink A B -Text text
                     }
 
                     It works {
@@ -689,7 +709,7 @@ flowchart LR
                 Context with-tooltip {
 
                     BeforeEach {
-                        $diagram | Add-MermaidFlowchartClick foo 'http://localhost' -Tooltip 'home sweet home' -Target blank
+                        $diagram | Add-MermaidFlowchartClick foo 'http://localhost' -Tooltip '<- home sweet home' -Target blank
                     }
 
                     It works {
@@ -697,7 +717,7 @@ flowchart LR
                         $output | Should -Not -BeNullOrEmpty
                         $output | Should -Be (@"
 flowchart LR
-    click foo "http://localhost" "home sweet home" _blank
+    click foo "http://localhost" "&lt;- home sweet home" _blank
 "@.Replace("`r`n", [Environment]::NewLine))
                     }
                 }
@@ -756,7 +776,7 @@ flowchart LR
                 Context nested-subgraphs {
 
                     BeforeEach {
-                        $subgraph | Add-MermaidFlowchartSubgraph bar 'this is a nested subgraph'
+                        $subgraph | Add-MermaidFlowchartSubgraph bar '<- this is a nested subgraph'
                     }
 
                     It works {
@@ -765,7 +785,7 @@ flowchart LR
                         $output | Should -Be (@"
 flowchart LR
     subgraph foo
-        subgraph bar [this is a nested subgraph]
+        subgraph bar [&lt;- this is a nested subgraph]
         end
     end
 "@.Replace("`r`n", [Environment]::NewLine))