+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Wrong PowerShellVersion in manifest file.
 - Added config to diagrams.
+- Escaped strings in flowcharts.
 
 ## [0.5.0] - 2023-09-22